x86emu: fixed warnings.

+++ source/drivers/x86/video/vbe.cc
@@ -189,7 +189,7 @@ uint16_t VBE::x86emuExec(uint16_t eax,uint16_t ebx,uint16_t ecx,uint16_t edi,uin
 EXTERN_C void vid_sprintf(char *str,const char *fmt,...) {
 	va_list ap;
 	va_start(ap,fmt);
-	vsnprintf(str,std::numeric_limits<size_t>::max(),fmt,ap);
+	vsnprintf(str,std::numeric_limits<int>::max(),fmt,ap);
 	va_end(ap);
 }
